Moving to the PEG parser.
