+++ alteruphono/__init__.py
@@ -6,7 +6,9 @@ __author__ = "Tiago Tresoldi"
 __email__ = "tresoldi@shh.mpg.de"
 
 # Build the namespace
-from alteruphono.sound_changer import apply_rule
+from alteruphono import compiler
+from alteruphono import convert
 from alteruphono import utils
 from alteruphono.grammar import SOUND_CHANGEParser as Parser
-from alteruphono import convert
+
+# from alteruphono.sound_changer import apply_rule
