+++ alteruphono/utils.py
@@ -7,12 +7,179 @@ from os import path
 import random
 import re
 
-# Import from namespace
-from . import sound_changer
+# Import 3rd party libraries
+from pyclts import CLTS
+
+_TRANSCRIPTION = CLTS().bipa
 
 # Set the resource directory; this is safe as we already added
 # `zip_safe=False` to setup.py
 _RESOURCE_DIR = path.join(path.dirname(path.dirname(__file__)), "resources")
+
+
+def parse_features(text):
+    """
+    Parse a list of feature constraints.
+
+    Constraints can be definied inside optional brackets. Features are
+    separated by commas, with optional spaces around them, and have a
+    leading plus or minus sign (defaulting to plus).
+
+    Parameters
+    ----------
+    text: string
+        A string with the feature constraints specification
+
+    Returns
+    -------
+    positive: list
+        A list of features to be included.
+    negative: list
+        A list of features to be excluded.
+    """
+
+    # Remove any brackets from the text that was received and strip it.
+    # This allows to generalize this function, so if that it can be used
+    # in different contexts (parsing both stuff as "[+fricative]" and
+    # "+fricative").
+    text = text.replace("[", "")
+    text = text.replace("]", "")
+    text = text.strip()
+
+    # Analyze all features and build a list of positive and negative
+    # features; if a feature is not annotated for positive or negative
+    # (i.e., no plus or minus sign), we default to positive.
+    positive = []
+    negative = []
+    for feature in text.split(","):
+        # Strip once more, as the user might add spaces next to the commas
+        feature = feature.strip()
+
+        # Obtain the positivity/negativity of the feature
+        if feature[0] == "-":
+            negative.append(feature[1:])
+        elif feature[0] == "+":
+            positive.append(feature[1:])
+        else:
+            positive.append(feature)
+
+    return positive, negative
+
+
+# NOTE: This function is used mostly by features2regex()
+def features2sounds(positive, negative, transsys):
+    """
+    Returns a list of graphemes matching positive and negative features.
+
+    Positive and negative are lists of features as defined in the
+    TranscriptionSystem.
+
+    For example, asking for not-rounded and not high front vowels:
+
+    >>> alteruphono.sound_changer.features2sounds(["vowel", "front"], ["rounded", "high"])
+    ['ḭːː', 'aa', 'ɛ̯', 'ĩĩ', 'ĕ', ... 'a˞', 'ẽ̞', 'iːː', 'e̯', 'aː', 'ii']
+
+    Parameters
+    ----------
+    positive : list
+        A list of strings with the features to be included.
+    negative : list
+        A list of strings with the features for be excluded.
+    transsys : TranscriptionSystem
+        The transcription system to be used.
+
+    Returns
+    -------
+    sounds : list
+        A list of strings with all the graphemes matching the requested
+        feature constraints.
+    """
+
+    # Iterate over all sounds in the transcription system
+    sounds = []
+    for sound in transsys.sounds:
+        # Extract all the features of the current sound
+        features = transsys[sound].name.split()
+
+        # Check if all positive features are there
+        pos_match = all(feat in features for feat in positive)
+
+        # Check if none of the negative features are there
+        neg_match = all(feat not in features for feat in negative)
+
+        # Accept the sound if it passes both tests
+        if pos_match and neg_match:
+            sounds.append(sound)
+
+    return sounds
+
+
+def features2regex(positive, negative, transsys=None):
+    """
+    Returns a regex string for matching graphemes according to features.
+
+    Positive and negative are lists of features as defined in the
+    TranscriptionSystem.
+
+    For example, asking for not-rounded and not high front vowels:
+
+    >>> alteruphono.sound_changer.features2regex(["vowel", "front"], ["rounded", "high"])
+    'ẽ̞ẽ̞|ẽ̞ː|ẽẽ|æ̃æ̃|ãã|a̰ːː|ḭːː|ĩĩ|ɛ̃ɛ̃|" ... "ă|e̤|ɛː|iː|aa|ɛ̯|ii|ḭ|æ|i|ɛ|e|a'
+
+    Parameters
+    ----------
+    positive : list
+        A list of strings with the features to be included.
+    negative : list
+        A list of strings with the features for be excluded.
+    transsys : TranscriptionSystem
+        The transcription system to be used, defaulting to BIPA.
+
+    Returns
+    -------
+    re_or_string : string
+        A string with the regular expression matching the requested feature
+        constraints. The items are sorted in inverse length order, and
+        include disjoints.
+    """
+
+    # Use the default transcription system, if none was provided
+    if not transsys:
+        transsys = _TRANSCRIPTION
+
+    # Get the list of sounds and sort it by inverse length, allowing the
+    # regular expression engine to correctly match them. There is no need
+    # for further sorting withing lenghs, such as alphabetical.
+    sounds = features2sounds(positive, negative, transsys)
+    sounds = sorted(sounds, key=len, reverse=True)
+
+    # Join all the sounds in a single regular expression string'; note that
+    # we *don't* add capturing parentheses here
+    re_or_string = "%s" % "|".join(sounds)
+
+    return re_or_string
+
+
+# TODO: remove hard-coding of fixes, loading internal or external data
+def fix_descriptors(descriptors):
+    # fix inconsistencies and problems with pyclts descriptors
+
+    # Run manual fixes related to pyclts
+    if "palatal" in descriptors and "fricative" in descriptors:
+        # Fricative palatals are described as alveolo-palatal, so
+        # replace all of them
+        descriptors = [
+            feature if feature != "palatal" else "alveolo-palatal"
+            for feature in descriptors
+        ]
+
+    if "alveolo-palatal" in descriptors and "fricative" in descriptors:
+        descriptors.append("sibilant")
+
+    if "alveolar" in descriptors and "fricative" in descriptors:
+        descriptors.append("sibilant")
+
+    return descriptors
 
 
 def read_sound_classes(filename=None):
@@ -38,9 +205,10 @@ def read_sound_classes(filename=None):
     with open(filename) as tsvfile:
         reader = csv.DictReader(tsvfile, delimiter="\t")
         sound_classes = {
-            row["sound_class"]: sound_changer.features2regex(
-                *sound_changer.parse_features(row["features"])
-            )
+            row["sound_class"]: {
+                "description": row["description"],
+                "regex": features2regex(*parse_features(row["features"])),
+            }
             for row in reader
         }
